Remove gulp and replace with webpack
