+++ package.json
@@ -4,7 +4,10 @@
   "description": "A web based game to teach students to read music.",
   "main": "dist/index.html",
   "scripts": {
-    "test": "test"
+    "build": "webpack",
+    "start:dev": "webpack-dev-server",
+    "start": "http-server dist",
+    "start:temp": "http-server src"
   },
   "repository": {
     "type": "git",
@@ -20,23 +23,19 @@
   },
   "homepage": "https://github.com/C0deJack/pianoApp#readme",
   "devDependencies": {
-    "browser-sync": "^2.26.3",
-    "colors": "^1.3.3",
-    "del": "^4.1.0",
-    "gulp": "^4.0.0",
-    "gulp-autoprefixer": "6.0.0",
-    "gulp-concat": "2.6.1",
-    "gulp-cssnano": "2.1.3",
-    "gulp-eslint": "^5.0.0",
-    "gulp-flatmap": "1.0.2",
-    "gulp-jshint": "2.1.0",
-    "gulp-optimize-js": "1.1.0",
-    "gulp-rename": "1.4.0",
-    "gulp-sass": "^4.0.2",
-    "gulp-svgmin": "2.1.0",
-    "gulp-terser": "1.1.7",
-    "jshint": "2.9.6",
-    "jshint-stylish": "2.2.1",
-    "lazypipe": "1.0.1"
+    "babel-core": "^6.26.3",
+    "babel-loader": "7.0.0",
+    "babel-preset-env": "^1.7.0",
+    "css-loader": "^2.1.1",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "http-server": "^0.11.1",
+    "mini-css-extract-plugin": "^0.6.0",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.1",
+    "webpack-dev-server": "^3.3.1"
   }
 }
