+++ src/index.html
@@ -3,8 +3,6 @@
 <head>
 	<meta charset="UTF-8">
 	<title>PIANO PWA</title>
-	<link rel="stylesheet" href="css/main.min.css">
-	<script src="js/play.min.js" defer></script>
 </head>
 <body>
 <div>
