+++ src/index.js
@@ -1,0 +1,6 @@
+import { play } from './js/play';
+import randomNoteGenerator from './js/libs/randomNoteGenerator';
+import './scss/main.scss'
+
+randomNoteGenerator();
+play();
